Перенёс код в модуль
